requirements.txt
flask==3.0.2
requests==2.31.0
neo4j==5.14.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-community==0.0.27
langchain-core==0.1.27
pydantic==2.6.3 